cleaning up long term evolution plots

+++ Ramses_analysis/FU_ori_investigation/long_term_evolution_multi.py
@@ -31,43 +31,51 @@ smoothing_window = yt.YTArray(100, 'yr')
 for sink_ind in sink_inds:
     pickle_file = '/groups/astro/rlk/rlk/FU_ori_investigation/Sink_pickles/particle_data_'+str(sink_ind)+'.pkl'
     if os.path.isfile(pickle_file):
-        print('reading ', pickle_file)
-        file_open = open(pickle_file, 'rb')
-        particle_data, counter, sink_ind, sink_form_time = pickle.load(file_open)
-        file_open.close()
-        print('Finished reading pickle, calculating smoothed quantities')
-        
-        mass_ratio = yt.YTArray(particle_data['mass']).T[0]/yt.YTArray(particle_data['mass']).T[1]
-        smooth_t = []
-        smooth_q = []
-        smooth_e = []
-        for it in range(len(particle_data['time'])):
-            curr_time = particle_data['time'][it]
-            start_time = particle_data['time'][it] - smoothing_window/2
-            if start_time < 0:
-                start_time = 0
-            start_it = np.argmin(abs(yt.YTArray(particle_data['time']) - start_time))
+        if os.path.isfile('/groups/astro/rlk/rlk/FU_ori_investigation/Sink_pickles/smoothed_particle_data_'+str(sink_ind)+'.pkl'):
+            file_open = open(pickle_file, 'rb')
+            smooth_t, smooth_q, smooth_e = pickle.load(file_open)
+            file_open.close()
             
-            end_time = particle_data['time'][it] + smoothing_window/2
-            if end_time > particle_data['time'][-1]:
-                end_time = particle_data['time'][-1]
-            end_it = np.argmin(abs(yt.YTArray(particle_data['time']) - end_time))
+            axs.flatten()[0].plot(smooth_t, smooth_q, alpha=0.25, label=str(sink_ind))
+            axs.flatten()[1].plot(smooth_t, smooth_e, alpha=0.25)
+        else:
+            print('reading ', pickle_file)
+            file_open = open(pickle_file, 'rb')
+            particle_data, counter, sink_ind, sink_form_time = pickle.load(file_open)
+            file_open.close()
+            print('Finished reading pickle, calculating smoothed quantities')
             
-            mean_t = np.mean(particle_data['time'][start_it:end_it])
-            mean_q = np.mean(mass_ratio[start_it:end_it])
-            mean_e = np.mean(particle_data['eccentricity'][start_it:end_it])
-            smooth_t.append(mean_t)
-            smooth_q.append(mean_q)
-            smooth_e.append(mean_e)
-        
-        axs.flatten()[0].plot(smooth_t, smooth_q, alpha=0.25, label=str(sink_ind))
-        axs.flatten()[1].plot(smooth_t, smooth_e, alpha=0.25)
-        
-        print('updating pickle')
-        file = open('/groups/astro/rlk/rlk/FU_ori_investigation/Sink_pickles/smoothed_particle_data_'+str(sink_ind)+'.pkl', 'wb')
-        pickle.dump((smooth_t, smooth_q, smooth_e), file)
-        file.close()
-        print('Finished updating pickle', '/groups/astro/rlk/rlk/FU_ori_investigation/Sink_pickles/smoothed_particle_data_'+str(sink_ind)+'.pkl')
+            mass_ratio = yt.YTArray(particle_data['mass']).T[0]/yt.YTArray(particle_data['mass']).T[1]
+            smooth_t = []
+            smooth_q = []
+            smooth_e = []
+            for it in range(len(particle_data['time'])):
+                curr_time = particle_data['time'][it]
+                start_time = particle_data['time'][it] - smoothing_window/2
+                if start_time < 0:
+                    start_time = 0
+                start_it = np.argmin(abs(yt.YTArray(particle_data['time']) - start_time))
+                
+                end_time = particle_data['time'][it] + smoothing_window/2
+                if end_time > particle_data['time'][-1]:
+                    end_time = particle_data['time'][-1]
+                end_it = np.argmin(abs(yt.YTArray(particle_data['time']) - end_time))
+                
+                mean_t = np.mean(particle_data['time'][start_it:end_it])
+                mean_q = np.mean(mass_ratio[start_it:end_it])
+                mean_e = np.mean(particle_data['eccentricity'][start_it:end_it])
+                smooth_t.append(mean_t)
+                smooth_q.append(mean_q)
+                smooth_e.append(mean_e)
+            
+            axs.flatten()[0].plot(smooth_t, smooth_q, alpha=0.25, label=str(sink_ind))
+            axs.flatten()[1].plot(smooth_t, smooth_e, alpha=0.25)
+            
+            print('updating pickle')
+            file = open('/groups/astro/rlk/rlk/FU_ori_investigation/Sink_pickles/smoothed_particle_data_'+str(sink_ind)+'.pkl', 'wb')
+            pickle.dump((smooth_t, smooth_q, smooth_e), file)
+            file.close()
+            print('Finished updating pickle', '/groups/astro/rlk/rlk/FU_ori_investigation/Sink_pickles/smoothed_particle_data_'+str(sink_ind)+'.pkl')
 
     axs.flatten()[0].set_xlim([0, 75000])
     axs.flatten()[0].set_ylabel('$q$ w.r.t closest sink')
